pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "pantra"
license  = "Apache-2.0"
authors = [
    {name = "Andrey Aseev", email = "andrey@aseev.biz"}
]
maintainers = [
    {name = "Andrey Aseev", email = "andrey@aseev.biz"}
]
urls = {Homepage = "https://github.com/zergos/pantra"}
description = "Python Full-stack Web Framework"
dependencies = [
	'uvicorn',
    'starlette',
    'websockets',
    'antlr4-python3-runtime==4.10',
    'cssutils>=2.3.0',
    'requests>=2.31.0',
    'libsass>=0.22.0',
    'Babel>=2.9.1',
    'watchdog>=2.1.6',
    'hsluv>=5.0.2',
]
requires-python = ">=3.10"
dynamic = ["version", "readme"]
scripts = {pantra = "pantra.management:execute_from_command_line"}

[project.optional-dependencies]
quazy = [
    'quazydb'
]

[tool.setuptools]
package-dir = {pantra = "pantra"}
include-package-data = true
package-data = {pantra = ["*.pyi", "js/*.js", "js/*.json"]}

[tool.setuptools.dynamic]
version = {attr = "pantra.VERSION"}
readme = {file = ["README.md"]}
